+++ app/Http/Controllers/OrderController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use stdClass;
 use App\Models\Cart;
 use App\Models\Order;
 use Illuminate\Http\Request;
@@ -11,60 +12,52 @@ use Illuminate\Support\Facades\Auth;
 class OrderController extends Controller
 {
     public function store(Request $request) {
+        $validator = $request->validate([
+            'location' => 'required|string|max:225',
+        ]);
+
         $userId = Auth::user()->id;
 
         $cartItems = DB::table('carts')
                         ->join('products', 'carts.product_id', 'products.id')
+                        ->select('carts.*', 'products.name', 'products.price')
                         ->where('user_id', $userId)
                         ->get();
 
-        foreach ($cartItems as $cartItem) {
-            Order::create([
-                'user_id' => $cartItem->user_id,
-                'product_id' => $cartItem->product_id,
-                'product_count' => $cartItem->product_count,
-                'location' => $request->location,
-                'total_cost' => $cartItem->price * $cartItem->product_count,
+        if(count($cartItems) > 0) {
+            foreach ($cartItems as $cartItem) {
+                Order::create([
+                    'user_id' => $cartItem->user_id,
+                    'product_id' => $cartItem->product_id,
+                    'product_count' => $cartItem->product_count,
+                    'location' => $request->location,
+                    'total_cost' => $cartItem->price * $cartItem->product_count,
+                ]);
+            }
+
+            Cart::where('user_id', $userId)->delete();
+
+            return response()->json([
+                'status' => 200,
+                'message' => 'Order successfully created.'
             ]);
+        } else {
+            return response()->json([
+                'status' => 422,
+                'message' => 'No cart items to order.'
+            ], 422);
         }
-
-        Cart::where('user_id', $userId)->delete();
-
-        return response()->json([
-            'status' => 200,
-            'message' => 'Order successfully created.'
-        ]);
     }
-
-
-    // public function store(Request $request) {
-    //     $fields = $request->validate([
-    //         'product_id' => 'required|integer|max:225',
-    //         'product_count' => 'required|integer|max:2',
-    //         'location' => 'required|string|max:225',
-    //         'total_cost' => 'required|string|max:20',
-    //     ]);
-
-    //     $userId = Auth::user()->id;
-
-    //     Order::create([
-    //         'user_id' => $userId,
-    //         'product_id' => $fields['product_id'],
-    //         'product_count' => $fields['product_count'],
-    //         'location' => $fields['location'],
-    //         'total_cost' => $fields['total_cost'],
-    //     ]);
-
-    //     return response()->json([
-    //         'status' => 201,
-    //         'message' => 'Order successfully created',
-    //     ], 201);
-    // }
 
 
     // list of all order
     public function list() {
-        $orderData = Order::paginate(8);
+        $orderData = Order::join('users', 'orders.user_id', 'users.id')
+                        ->join('products', 'orders.product_id', 'products.id')
+                        ->select('orders.*', 'users.name as user_name', 'users.email', 'users.phone',
+                        'users.address', 'users.image as user_image', 'products.name as product_name',
+                        'products.price', 'products.image as product_image')
+                        ->paginate(8);
 
         return response($orderData);
     }
@@ -73,10 +66,12 @@ class OrderController extends Controller
     public function orderListOfAUser() {
         $userId = Auth::user()->id;
 
-        $orderData = DB::table('orders')
+        $orderData = Order::join('users', 'orders.user_id', 'users.id')
                         ->join('products', 'orders.product_id', 'products.id')
-                        ->select('orders.*', 'products.name')
-                        ->where('user_id', $userId)
+                        ->select('orders.*', 'users.name as user_name', 'users.email', 'users.phone',
+                        'users.address', 'users.image as user_image', 'products.name as product_name',
+                        'products.price', 'products.image as product_image')
+                        ->where('orders.user_id', $userId)
                         ->get();
 
         return response($orderData);
@@ -84,35 +79,72 @@ class OrderController extends Controller
 
     // Order accept
     public function accept($orderId) {
-        Order::where('id', $orderId)->update([
-            'status' => 'accepted',
-        ]);
+        $resp = $this->OrderActions($orderId, 'accept');
 
         return response()->json([
-            'status' => 200,
-            'message' => 'Order successfully accepted',
-        ]);
+            'status' => $resp->status,
+            'message' => $resp->mesg,
+        ], $resp->status);
     }
 
     // Order rject
     public function reject($orderId) {
-        Order::where('id', $orderId)->update([
-            'status' => 'rejected',
-        ]);
+        $resp = $this->OrderActions($orderId, 'reject');
 
         return response()->json([
-            'status' => 200,
-            'message' => 'Order successfully rejected',
-        ]);
+            'status' => $resp->status,
+            'message' => $resp->mesg,
+        ], $resp->status);
     }
 
     // Order delete
     public function destory($orderId) {
-        Order::where('id', $id)->delete();
+        $resp = $this->OrderActions($orderId, 'delete');
 
         return response()->json([
-            'status' => 200,
-            'message' => 'Order successfully deleted',
-        ]);
+            'status' => $resp->status,
+            'message' => $resp->mesg,
+        ], $resp->status);
+    }
+
+    // Order action handler
+    private function OrderActions($id, $action) {
+        $order = Order::where('id', $id)->get()->first();
+        $resp = new stdClass();
+
+        if($order != null) {
+            switch ($action) {
+                case 'accept':
+                    Order::where('id', $orderId)->update([
+                        'status' => 'accepted',
+                    ]);
+                    $resp->status = 200;
+                    $resp->mesg = 'Order successfully accepted.';
+                    break;
+
+                case 'reject':
+                    Order::where('id', $orderId)->update([
+                        'status' => 'rejected',
+                    ]);
+                    $resp->status = 200;
+                    $resp->mesg = 'Order successfully rejected.';
+                    break;
+
+                case 'delete':
+                    Order::where('id', $id)->delete();
+                    $resp->status = 200;
+                    $resp->mesg = 'Order successfully deleted.';
+                    break;
+
+                default:
+                    # code...
+                    break;
+            }
+        } else {
+            $resp->status = 422;
+            $resp->mesg = 'No item to perform actions';
+        }
+
+        return $resp;
     }
 }
